+++ packages/react-libraries/lib/generateReadme.js
@@ -3,13 +3,14 @@
 const fs = require('fs-extra');
 const get = require('lodash.get');
 const logger = require('./logger');
-const generateHeader = require('./generateHeader');
+const headers = require('./headers');
+const canGenerateFile = require('./canGenerateFile');
 
 const dst = 'README.md';
 
 function generateReadme(src = 'README-template.md') {
-  if (fs.existsSync(src)) {
-    const header = generateHeader(src);
+  if (fs.existsSync(src) && canGenerateFile(dst)) {
+    const header = headers.generateHeader(src);
     const pkg = fs.readJSONSync('package.json');
 
     const readme = fs
@@ -22,13 +23,6 @@ function generateReadme(src = 'README-template.md') {
         const value = get(pkg, path, null);
         return `${escapeChar}${value}` || match;
       });
-
-    // // Replace all unescaped variables.
-    // .replace(/([^\\])\$\{VERSION\}/g, `$1${pkg.version}`)
-    // .replace(/([^\\])\$\{DESCRIPTION\}/g, `$1${pkg.description}`)
-    // .replace(/([^\\])\$\{NAME\}/g, `$1${pkg.name}`)
-    // // Remove all `\` from escaped variables (ex: `\${VERSION}` => `${VERSION}` )
-    // .replace(/\\(\$\{(VERSION|DESCRIPTION|NAME)\})/g, '$1');
 
     fs.writeFileSync(dst, `${header}${readme}`);
     logger.generated(src, dst);